+++ Models/XygeniIssue.cs
@@ -120,6 +120,8 @@ namespace vs2026_plugin.Models
         public abstract string GetIssueDetailsHtml();
         public abstract string GetCodeSnippetHtmlTab();
 
+        public virtual string GetCodeFlowTab() => "";
+
         public virtual string GetSubtitleLineHtml()
         {
             string subtitle = CategoryName;